Fix another potential bug [v0.29.12]

+++ api/src/main/java/com/ryderbelserion/fusion/api/utils/FileUtils.java
@@ -119,8 +119,10 @@ public class FileUtils {
                     continue;
                 }
 
-                try (final InputStream stream = jar.getInputStream(entry)) {
-                    Files.copy(stream, path);
+                if (!Files.exists(path)) {
+                    try (final InputStream stream = jar.getInputStream(entry)) {
+                        Files.copy(stream, path);
+                    }
                 }
 
                 processedFiles.add(entryName);
